Rename a few map indexing classes
Rename classes:
  MapIndexer -> MapIndexingTask
  MapIndexingTask -> MapIndexingTaskRunner
  MapIndexResult -> MapIndexingResult
