+++ spitfire-server/maps-module/src/test/resources/logback-test.xml
@@ -9,5 +9,5 @@
         <appender-ref ref="STDOUT"/>
     </root>
 
-    <logger name="org.triplea.maps.indexing.MapIndexer" level="OFF"/>
+    <logger name="org.triplea.maps.indexing.MapIndexingTask" level="OFF"/>
 </configuration>
